Reorder some chapters
Functions are now the 2nd chapter following the Basics in Part 1.
Getting in and out begin the Part 2.
One exercise in the former last chapter about Functions was removed (redundant about if-statement).
Some exercises were moved between chapters as they didn't belong to their former chapters.
A missing exercise section was added to 'Getting data in and out'